Extract Prompt Studio settings helpers
