+++ web/js/easyuse_anima_prompt_studio.js
@@ -1,5 +1,5 @@
 import { app } from "../../../scripts/app.js";
-import { easyuseAnimaClassifyPrompt, easyuseAnimaGetSettings } from "./easyuse_anima_api.js";
+import { easyuseAnimaClassifyPrompt } from "./easyuse_anima_api.js";
 import { easyuseAnimaText, easyuseAnimaWatchLocale } from "./easyuse_anima_i18n.js";
 import { normalizePromptTagText } from "./easyuse_anima_prompt_rules.js";
 import {
@@ -18,10 +18,6 @@ import {
   LEGEND_ROW_HEIGHT,
   LEGEND_COLUMNS,
   STUDIO_WIDGET_VERTICAL_GAP,
-  PROMPT_STUDIO_FONT_SIZE_DEFAULT,
-  PROMPT_STUDIO_FONT_SIZE_MIN,
-  PROMPT_STUDIO_FONT_SIZE_MAX,
-  PROMPT_STUDIO_FONT_FAMILY,
   WEIGHT_NUMBER_RE,
   WEIGHTED_TOKEN_RE,
   WEIGHT_NUMBER_COLOR,
@@ -48,11 +44,6 @@ import {
 } from "./prompt_studio/constants.js";
 import {
   debounce,
-  isHexColor,
-  hexToRgba,
-  parseColorSettings,
-  normalizePromptStudioFontSize,
-  normalizePromptStudioFontFamily,
   escapeHtml,
   escapeAttr,
   advancedResolutionLabel,
@@ -101,6 +92,12 @@ import {
   ensureAdvancedStyle,
   ensureExtendSlotStyle,
 } from "./prompt_studio/style.js";
+import {
+  PROMPT_STUDIO_SETTINGS,
+  applyPromptStudioSettings,
+  applyPromptStudioTextStyle,
+  loadPromptStudioSettings,
+} from "./prompt_studio/settings.js";
 import {
   advancedEditorMinimumHeight,
   advancedEditorWidgetHeight,
@@ -152,16 +149,6 @@ function sectionLabel(section) {
   return psText(`section.${key}`) || style?.label || key;
 }
 
-const PROMPT_STUDIO_SETTINGS = {
-  typoIndicator: true,
-  weightSyntaxUnderline: false,
-  commentItalic: true,
-  fontOverride: false,
-  fontFamily: "",
-  fontSize: PROMPT_STUDIO_FONT_SIZE_DEFAULT,
-  trainedTagTooltip: true,
-  naiaGeneralAboveAutoToggle: false,
-};
 let middlePanForwardActive = false;
 let promptStudioTagTooltip = null;
 let promptStudioTagTooltipMoveFrame = 0;
@@ -462,82 +449,6 @@ function refreshNodeSize(node, options = {}) {
 
 
 
-
-function applyPromptStudioTextStyle(input) {
-  if (!(input instanceof HTMLElement)) {
-    return;
-  }
-  const nextFamily = (PROMPT_STUDIO_SETTINGS.fontOverride && PROMPT_STUDIO_SETTINGS.fontFamily)
-    ? PROMPT_STUDIO_SETTINGS.fontFamily
-    : PROMPT_STUDIO_FONT_FAMILY;
-  if (input.style.fontFamily !== nextFamily) {
-    input.style.fontFamily = nextFamily;
-  }
-  if (!PROMPT_STUDIO_SETTINGS.fontOverride) {
-    if (input.dataset.easyuseAnimaPromptTextOverride === "true") {
-      input.style.fontSize = "";
-      delete input.dataset.easyuseAnimaPromptTextOverride;
-    }
-    return;
-  }
-  const nextSize = `${PROMPT_STUDIO_SETTINGS.fontSize}px`;
-  if (input.style.fontSize !== nextSize) {
-    input.style.fontSize = nextSize;
-  }
-  input.dataset.easyuseAnimaPromptTextOverride = "true";
-}
-
-function applyPromptStudioSettings(settings) {
-  PROMPT_STUDIO_SETTINGS.typoIndicator = settings?.["prompt_studio.typo_indicator"] !== "false";
-  PROMPT_STUDIO_SETTINGS.weightSyntaxUnderline = settings?.["prompt_studio.weight_syntax_underline"] === "true";
-  PROMPT_STUDIO_SETTINGS.commentItalic = settings?.["prompt_studio.comment_italic"] !== "false";
-  PROMPT_STUDIO_SETTINGS.trainedTagTooltip = settings?.["prompt_studio.trained_tag_tooltip"] !== "false";
-  if (!PROMPT_STUDIO_SETTINGS.trainedTagTooltip) {
-    hideTrainedTagTooltip();
-  }
-  PROMPT_STUDIO_SETTINGS.fontOverride = settings?.["prompt_studio.font_override"] === "true";
-  PROMPT_STUDIO_SETTINGS.fontFamily = normalizePromptStudioFontFamily(settings?.["prompt_studio.font_family"]);
-  PROMPT_STUDIO_SETTINGS.fontSize = normalizePromptStudioFontSize(settings?.["prompt_studio.font_size"]);
-  if (PROMPT_STUDIO_SETTINGS.fontOverride) {
-    document.documentElement?.style?.setProperty(
-      "--easyuse-anima-prompt-studio-font-size",
-      `${PROMPT_STUDIO_SETTINGS.fontSize}px`,
-    );
-    document.documentElement?.style?.setProperty(
-      "--easyuse-anima-prompt-studio-font-family",
-      PROMPT_STUDIO_SETTINGS.fontFamily || PROMPT_STUDIO_FONT_FAMILY,
-    );
-  } else {
-    document.documentElement?.style?.removeProperty("--easyuse-anima-prompt-studio-font-size");
-    document.documentElement?.style?.removeProperty("--easyuse-anima-prompt-studio-font-family");
-  }
-  PROMPT_STUDIO_SETTINGS.naiaGeneralAboveAutoToggle =
-    settings?.["prompt_studio.naia_general_above_auto_toggle"] === "true";
-  const colors = parseColorSettings(settings?.["prompt_studio.colors"]);
-  for (const [key, color] of Object.entries(colors)) {
-    if (!SECTION_STYLES[key] || !isHexColor(color)) {
-      continue;
-    }
-    SECTION_STYLES[key].color = color;
-    if (SECTION_STYLES[key].background && SECTION_STYLES[key].background !== "transparent") {
-      SECTION_STYLES[key].background = hexToRgba(color, 0.18);
-    }
-  }
-}
-
-async function loadPromptStudioSettings() {
-  try {
-    const settings = await easyuseAnimaGetSettings({ fallback: null });
-    if (!settings) {
-      return;
-    }
-    applyPromptStudioSettings(settings);
-    refreshAllPromptHighlights(true);
-    app.graph?.setDirtyCanvas(true, true);
-  } catch {
-    // Keep built-in defaults if the settings endpoint is not available yet.
-  }
-}
 
 async function classifyPrompt(text) {
   return easyuseAnimaClassifyPrompt(text);
@@ -4355,7 +4266,13 @@ app.registerExtension({
     installMiddlePanForwarder();
     installAdvancedSaveSync(app, syncAllAdvancedNodes);
     installPromptHighlightOverlayRefresh();
-    await loadPromptStudioSettings();
+    await loadPromptStudioSettings({
+      hideTrainedTagTooltip,
+      afterApply: () => {
+        refreshAllPromptHighlights(true);
+        app.graph?.setDirtyCanvas(true, true);
+      },
+    });
     easyuseAnimaWatchLocale(() => {
       refreshPromptStudioLocaleDom();
       refreshAllPromptHighlights();
@@ -4364,7 +4281,7 @@ app.registerExtension({
       if (!event?.detail) {
         return;
       }
-      applyPromptStudioSettings(event.detail);
+      applyPromptStudioSettings(event.detail, { hideTrainedTagTooltip });
       for (const node of app.graph?._nodes || []) {
         if (isAdvancedNode(node)) {
           renderAdvancedEditor(node);
